daemon: common pop_blocks impl
